Pulled object timeline code into a separate tools script
